+++ src/ga4/Analytics.php
@@ -297,8 +297,18 @@ class Analytics
     {
         if ($this->_request === null) {
             $this->_request = new BaseRequest();
+
             $this->_request->setClientId(AnalyticsHelper::getClientId());
+
+            if (InstantAnalytics::$settings->sendUserId) {
+                $userId = AnalyticsHelper::getUserId();
+
+                if ($userId) {
+                    $this->request()->setUserId($userId);
+                }
+            }
         }
+
 
         return $this->_request;
     }