+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 ## 4.0.0-beta.3 - UNRELEASED
 ## Added
 * Added the `sessionDuration` setting that can be changed via config.php.
+* Added the `instantanalytics.beginCheckout(cart)` action.
+* Added the `sendUserId` setting that can be changed via config.php
 
 ### Changed
 * `addCommerceProductImpression()` no longer supports list index and list name. ([#6](https://github.com/nystudio107/craft-instantanalytics-ga4/issues/6))
+* Instant Analytics GA4 now supports sending user id for logged in users automatically.
 
 ### Fixed
 * Fixed an issue where other plugin status was checked too early.